requirements.txt
torch
transformers
accelerate
sentence-transformers
numpy
scikit-learn
matplotlib
seaborn
tqdm
huggingface_hub

# API clients for fast prompt generation (optional)
groq
cerebras-cloud-sdk
openai
